Heavy Weapons Update (and knife)
Heavy weapons have been added to drop tables, and have a chance of replacing BFG spawns. This is rarer than the Gungnir and other BFG replacements due to the weapons' raw power comparatively.

Aside from this, any belt box spawns now include a pile of belt links, so as to reduce the amount of brass-and-link double-goblin'ing you have to do.

Lastly, the knife has been moved in the slot selection, and is now selected before your fists, giving it a quick-draw advantage that the other melee weapons lack. It's also received a tiny range increase (of 6 units), allowing it to reach things that your fist would otherwise JUST barely miss.