Melee shit to keep up-to-date
The code for kick/lunge got changed up, and the melee weapons referenced the old code, so I ported the old text block from an older version of the mod where it was contained in the zsc for the melee weapons, and they work again. Exactly the same as before, so they lunge differently from fists now.

There's also some terminator changes that have thus-far been unsuccessful in fixing the issue with getting the custom pin to display properly, and literally nothing I do works and I have zero idea why.