First Pass of .50AM "rework"
with a good bit of help and asking around I went and started on un-fucking the ballistics on 50AM, massively reducing it's absurd mass of 1563 down to 260, along with tweaking the rest of the stats to bring it more in line with bulletlib's 50AE. It's still a stronger round, but it isn't going to instagib imps and zombiemen anymore.

On a similar note, I reduced the speed multiplier on the Lever-action Rifle from 1.5 to 1.2, to keep it from STILL instagibbing things like that. This is only a bandaid fix however, and the stats will probably change more over time.

Note, I *still* haven't gotten around to fixing the fire extinguisher for freedoom. That's next on the list.