FreeDoom Fire Extinguisher and slight melee tweaks
The mod now works in freedoom because I got off my ass and added freedoom filter files for the extinguisher thanks to it not having lost souls.

Also I nerfed the axe's geometry destruction power a bit in a way I don't quite understand. Much like the .50AM changes, it'll be fine-tuned a bit as I figure out how it works more.

Also the sledgehammer now has 10 more bulk, putting it at 75 bulk, compared to the axe's 55 and the chainsaw's 100(plus battery).