teeny-tiny update to the suppressed 4mm bullet type used by the g11
I completely forgot the silenced G11 fires a special kind of 4mm that doesn't have the far-distance sound effect, and was reminded by people saying that the g11s had artificially inflated damage.

So I went to check and sure enough the silenced one did, but not for the reason I thought. Was actually just the speed not being updated to match current 4mm's, and nothing actually with the gun itself.